+++ src/ScientificDetectors.jl
@@ -45,9 +45,9 @@ import Base: read, write
 using Statistics, StatsBase, Distributions
 using MultivariateOnlineStatistics
 using ArrayTools, StructuredArrays
-using EasyFITS
-using EasyFITS: hduname, throw_file_already_exists, FitsImageHDU
-import EasyFITS: write!, readfits, writefits, writefits!
+using AstroFITS
+using AstroFITS: hduname, throw_file_already_exists, FitsImageHDU
+import AstroFITS: write!, readfits, writefits, writefits!
 
 function process end
 function process! end